+++ AutoSaver/_Thunderstore/CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.2.0] - 2024-Dec-17
+## [1.3.0] - 2024-May-22
+
+### Added
+
+- AutoSaver will now also save latest save files for the current game version
+  - These save files will not get touched when updating to a newer version, allowing you to restore / return to a previous version's save file if needed 
+
+## [1.2.0] - 2024-May-05
 
 ### Added